requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pandas>=2.1.0
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0
requests>=2.31.0

# ML and data processing
mlflow>=2.8.0
boto3>=1.34.0
s3fs>=2024.1.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Visualization and dashboard
streamlit>=1.28.0
plotly>=5.17.0
folium>=0.15.0

# Orchestration
prefect>=2.14.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.11.0
flake8>=6.1.0
isort>=5.12.0

# Monitoring
prometheus-client>=0.19.0

# Security
cryptography>=41.0.0

# Utilities
python-dotenv>=1.0.0
pytz>=2023.3
